pyproject.toml
[project]
name = "scheduler"
version = "0.1.0"
description = ""
authors = [
    {name = "Serhii Ovsiuk",email = "s.ovsiuk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "aiogram (>=3.25.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "redis (>=7.2.1,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"